Fix: All camera test

+++ includes/camera.h
@@ -6,7 +6,7 @@
 /*   By: njooris <njooris@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/09 12:09:03 by njooris           #+#    #+#             */
-/*   Updated: 2025/10/09 14:35:14 by njooris          ###   ########.fr       */
+/*   Updated: 2025/10/09 15:18:53 by dernst           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@
 typedef struct s_camera
 {
 	t_matrix4	transform;
-	uint32_t	vsize;
-	uint32_t	hsize;
+	double		vsize;
+	double		hsize;
 	double		half_width;
 	double		half_height;
 	double		fov;
@@ -32,7 +32,7 @@ typedef struct s_camera
 } t_camera;
 
 t_matrix4_ptr	view_transform(t_tuple from, t_tuple to, t_tuple up, t_matrix4 r);
-t_camera		camera(uint32_t hsize, uint32_t vsize, double fov);
+t_camera		camera(double hsize, double vsize, double fov);
 t_ray			ray_for_pixel(t_camera c, uint32_t px, uint32_t py);
 t_canvas		render(t_camera c, t_world w, t_canvas img);
